refactor(ast): consolidate utils into the src/utils folder
Move the internal AST utilities from src/utils.ts into src/utils/ast.ts so all
utils live under src/utils (alongside the public src/utils/index.ts that backs
the @kubb/ast/utils subpath). No API change.
