+++ packages/ast/src/utils/ast.ts
@@ -1,8 +1,8 @@
 import { camelCase, isIdentifier, isValidVarName, memoize, singleQuote } from '@internals/utils'
 
-import { INDENT } from './constants.ts'
-import { createFunctionParameter, createFunctionParameters, createParameterGroup, createParamsType, createProperty, createSchema } from './factory.ts'
-import { narrowSchema } from './guards.ts'
+import { INDENT } from '../constants.ts'
+import { createFunctionParameter, createFunctionParameters, createParameterGroup, createParamsType, createProperty, createSchema } from '../factory.ts'
+import { narrowSchema } from '../guards.ts'
 import type {
   CodeNode,
   ExportNode,
@@ -15,10 +15,10 @@ import type {
   ParamsTypeNode,
   SchemaNode,
   SourceNode,
-} from './nodes/index.ts'
-import type { SchemaType } from './nodes/schema.ts'
-import { extractRefName } from './refs.ts'
-import { collect, collectLazy } from './visitor.ts'
+} from '../nodes/index.ts'
+import type { SchemaType } from '../nodes/schema.ts'
+import { extractRefName } from '../refs.ts'
+import { collect, collectLazy } from '../visitor.ts'
 
 const plainStringTypes = new Set<SchemaType>(['string', 'uuid', 'email', 'url', 'datetime'] as const)
